Add versioning to IPC
Versioning is implemented as a raw u8 sent as a header from the
broker to its workers, before anything else is serialized.
We could also have implemented a first VersionMessage serialized
by bincode, but this would be a bit overkill and would require
future versions to embark bincode. This way, if we want to drop
bincode in the future, we can.