Add seccomp-unotify code to send file descriptors
Just used in open/openat handlers for now, but will also
be required for accept, etc.

/!\ BEWARE: this piece of code does not handle the race
condition which can occur between the time a file descriptor
is sent, and the time we call seccomp_notify_respond().
There is a chance we leak file descriptors in workers for now,
if they receive a signal in this time frame.