Split off clone() entrypoint into a nostd crate
This removes probabilistic deadlocks when using a stdlib feature
which uses a lock (e.g. memory allocation, or buffered I/O when
logging messages), when the lock is held by a thread which dies
during clone().

clone() is intrinsically unsafe to use in Rust without this, and
this is probably the best we can do.