Add thread synchronization when applying seccomp filter
When not using seccomp user notifications, there is no guarantee
that there isn't already other thread(s) created by the developper
or one of the libraries they use. These threads cannot be allowed
to run freely without a seccomp filter.

Thanks to @asenat for pointing this out! This commit should
fix #7 and add a regression test.