[linux-entrypoint] Make warning macro safe and usable in modules
The log_nonfatal!() macro now uses a global to store the configured
output file descriptor given by the broker. This will allow using
it in other modules than in the main() of the entrypoint (e.g.
capabilities)

Also make that macro safe, so that the entirety of the
entrypoint does not have to be unsafe, and so we avoid
warnings about "unused unsafes" when unsafe blocks get nested
in macro expansion.