[linux] Add newfstatat alongside fstat in seccomp
We can consider them to be equivalent. Ideally they should be filtered
to only allow paths which are readable or writable (or both). This
patch allows newfstatat unconditionnally, which allows e.g. leaking
file existence and size. Just a quick fix because tests now fail
because the rust standard library started using newfstatat.