[linux] Remove logging from unsafe signal handler
The real fix is going to be to rewrite this part of the code
in #[no_std] : even if the handler uses only syscalls that are
allowed without generating a second signal, it still needs to
not use locks that might be held by what is being preempted
(e.g. stdout locks in stdlib).