Move formatting stackbuffer to the ipc crate
IPC in Windows needs this primitive too to format
pipe names in no_std environment.

The disadvantage is it creates a dependency
from linux-entrypoint onto ipc, but it will be created
in the next few commits anyway when we migrate to seccomp
with user notifications, which will require sending IPC
messages post-fork-pre-exec in linux-entrypoint.